requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL driver
PyMySQL==1.1.0  # MySQL driver

# Environment & Configuration
python-dotenv==1.0.0

# Web & HTTP
requests==2.31.0

# Blockchain & Web3
web3==6.11.1
eth-account==0.10.0
eth-keys==0.4.0
eth-utils==2.1.0
eth-typing==3.5.0

# Production Server
gunicorn==21.2.0

# Error Tracking & Monitoring (Optional)
sentry-sdk==1.40.0

# Logging
python-json-logger==2.0.7

# Rate Limiting (Optional)
Flask-Limiter==3.5.0

# CORS & Security
cryptography==41.0.7